+++ src/AppBundle/Resources/views/layout.html.twig
@@ -29,7 +29,7 @@
 
     {% block body %}{% endblock %}
 
-    {{ render(controller('AppBundle:Nav:modalContact')) }}
+    {{ render(controller('AppBundle:Front:modalContact')) }}
     {{ include("AppBundle:Front:footer.html.twig") }}
